requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
cron-descriptor==1.4.5
cycler==0.12.1
Deprecated==1.2.18
distro==1.9.0
Django==5.0.2
django-celery-beat==2.7.0
django-debug-toolbar==4.3.0
django-filter==23.5
django-js-asset==2.2.0
django-mathfilters==1.0.0
django-mptt==0.16.0
django-phonenumber-field==7.3.0
django-timezone-field==7.1
et_xmlfile==2.0.0
flake8==7.0.0
fonttools==4.57.0
gunicorn==23.0.0
idna==3.8
iniconfig==2.1.0
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.4.1
matplotlib==3.10.1
mccabe==0.7.0
mysql-connector-python==8.3.0
mysqlclient==2.2.4
netaddr==1.3.0
numpy==2.2.4
openpyxl==3.1.5
packaging==25.0
phonenumbers==8.13.32
pillow==10.2.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.10
pycodestyle==2.11.1
pyflakes==3.2.0
pyparsing==3.2.3
pytest==8.3.5
pytest-django==4.11.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
redis==5.0.8
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.6.0
tzdata==2023.4
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
yookassa==3.5.0
